backend/pyproject.toml
[tool.poetry]
name = "jellylist"
version = "0.1.0"
description = ""
authors = ["Mohamed Raffik <raffikmohamed7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-ninja = "^1.2.0"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
